chapter_5/basic_master_slave_replication/requirements.txt
django==5.0.2
django-ninja==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1 